Rename applePay hook to useApplePay
